move profile.ts to Models
